Déplacer les affaires bloquées sur la page de revue
Le panneau vivait sur « Stats liaison », que personne n'ouvre pour trouver du travail,
et la page de revue n'y renvoyait pas.
